Change prompt string.
Using iTerm's shell integration, some padding was needed (the
integration show an indicator that shows if a command failed or
succeeded).
The " $ " in the beginning of the typing part of the prompt is just
taking up space and since I just learned that there is a visual back-
and forward keyboard shortcut to highlight through the history of
prompts, this is really no longer needed.