Set $EDITOR to the standard vim editor.
MacVim installation via homebrew had issues with my local ~/bin
symlinks.  They were removed in [2873e9f] and so the environment
variable had to change to the default vim command, not the symlink in
~/bin/  since the mvim version there was not working correctly anymore.

I have tried overriden my mvim with a newer one from the MacVim
installation, but that didn't work either.  So we'll fall back to the
standard vim install on OSX for the moment, should be enough for basic
stuff and git commit messages.