Git status is easily removed from prompt.
When dealing with large repositories (ex: thousands (15GB) of static
html pages); the prompt takes seconds to appear due to the git status in
the prompt.
This change involves 2 functions; one to remove the git status from the
prompt and one to restore the prompt.
It's a bit of quick work around, but because I don't need it very often
(and usually in a different window), I consider this "good enough"™

Closes #7