Add little extra to distinguish between command typed and its output.
Have been somewhat confused between the command I typed and its output
in that I thought that the command was actually the first line of the
output.  For some reason the blue clock part above the command line does
not register in some cases.  My brain still sees that blue part as the
line where the command was typed, and not the line under it.
By adding a little '-> ' this seem to trigger my mind to distinguish...

Also replaced the _no_git prompt with _simple prompt... we'll see how
that goes, I'm just trying it out.

And I added some dependency information to indicate where the colors
come from.