Put status bar back at the top.
In [81e83f8] I've moved it to the bottom to be less in-your-face, but it
takes up a few pixels more in height.
But the main reason to put it back to the top is that is looks less
messy when using Vim in the shell, since Vim _also_ has a status bar at
the bottom.  And I'm not willing to change Vim's status bar.