Prompt: simplify, while still providing enough information.
By making the usage of colors more dependent on the color scheme of the
terminal, switching color schemes will become easier and less intrusive.

By not changing the color of the prompt when the previous command exited
with an error code, the prompt remains consistent and thus easier to
distinguish while mouse scrolling through the history in the terminal
window.

By verbatim putting the exit code of the previous command in the prompt,
we can distinguish between different errors of the same command and not
just "fail/success".