bin/ym: Treat args as 1 file name.
When dealing with a file with spaces, bin/ym expect you to put it in
quotes so it can handle it as a single argument.  This is cumbersome to
deal with when using the `ym` alias, because that will only pass the
part of the file name up until the first space.

This does not consider special characters (for the shell), but a space
is about as far as I go in naming files ;-)