Git config: Better `lastversion` alias.
The previous version of the 'lastversion' alias would show older tags if
they were prefixed with 'v' (for instance).  If use the describe
version, this prefix is not taken into account, yielding correcter
results.
This is, however dependent on the branch you are, so executing this on
master is paramount to getting the correct results.