Bash prompt: align typing line with info line.
When executing a command, when glancing at the output, the command
sometimes is seen as part of the output.  When you do an `ls` for
instance, the "ls" seems to be a file that's part of the output of `ls`.
By aligning the prompt line with the info line in the same reverse
display, the distinction is pretty clear.  I actually don't think that
copypasting will be that more difficult (triple click).