Git config: better formatted release annotations.
When requesting a changelog, the output was somewhat messy.  In this
change we put the commit hash and subject on one line, the hash being
formatted in a GitHub-compatible manner.  And we indent with a pretty
standard 4 spaces, compensating for the indent in the formatting so that
the original commit message does not get formatted differently (Given
that one should wrap commit messages at 72 characters.)

The short tag message ("subject") is also changed to a more descriptive
"Release X" instead of directly listing all the changes (these come now
right after the initial tag message).