iTerm2: some more mocking about with cosmetics.
Changed components on status bar: current directory and git branch are
not really necessary since they are in my prompt.
The time is nice when in full screen mode (where it does not show the
mac osx clock) and now I can still see the bandwidth usage (which was
pushed out of sight when dealing with a terminal on half a screen).
This change also doesn't need separate status bars per pane anymore.