Restore normal prompt, iTerm2 status bar not good enough.
When checking out a tag, iTerm2's status bar shows the commit, not the
tag name; this confusing when checking out certain versions of a
project, only to see a piece of a sha1 hash.

Second thing (that I forgot about); you can not scroll back in history
and easily find out in which directory you executed a command.

The status bar is now of no real use anymore, so I've disabled it
altogether.