Make ssh_drush() more robust.
In [f365b60] some functionality was added to execute drush commands over
SSH.  `$sshuser` was totally wrong variable naming; it is in fact the
name of the application _and_ that application name is maximum 14 chars
long.  Thanks to @Jelle-S for pointing that out.

Also, the default `sed` command on OSX is FreeBSD's, which is strictly
POSIX compliant and different from GNU-sed, therefore we check for the
existence of `gsed` which should be installed.