iTerm2: Never use thin strokes.
In [1d50e24] we switched to a extra light font weight for IBM Plex Mono
for better readability, but it became obvious on the Mac's retina
display that this was too much with the 'Thin Strokes' setting
activated.  Without thin strokes, the font is nice again like on the
external monitor.