Replace `~` with `$HOME`
I was experimenting with custom Git command in `~/bin/` but it was not
being picked up by the git command.  By using `$HOME`, which gets
expanded to the absolute path of the home directory instead of remaining
`~` (which has different meaning to the `git` command), making Git pick
up the custom command as it should.