Add alias for MacVim.
'mvim' is just too cumbersome to type; it is mostly right hand-only.
Plus, I type it a lot during the day.  'e' (for 'edit') is short and a
decent mnemonic.
I am not caring about Vim/GVim, since MacVim is what I usually use,
shell Vim pops up when needed due to it being the $EDITOR.