ssh_load_autocomplete includes external Digipolis ssh config.
Domainator has the functionality to export SSH config entries for the
hosts.  For now it only supports a piece of text that you can copy
paste.  This is cumbersome in a manually crafted SSH config file.
Apparently we can use an `Include` statement in our local .ssh/config,
so we can just we can just override that included file when necessary,
making updating the ssh hosts from Domainator a tad more 'no brainer'.

Since `ssh_load_autocomplete` uses `awk` to parse the `.ssh/config`
file, we need to expand this to include the separate file also.  This is
what this change does.

Further automation is, however, advised so that the SSH config can be
kept up to date as painlessly as possible.
We also might need to extract the hardcoded `{config,digipolis-hosts}`
construct and put it in the .bash/config?