Add default FZF preview mode that supports tmux.
In [dbb865c] we've added preview to some FZF constructs, but FZF also
supports tmux if available (and if you are in a tmux session) to render
a nicer windowed preview.  Apparently, you can substitute `fzf` for
`fzf-tmux` without side effects because extra options that are only
available when in a tmux session, will be ignored when not in a tmux
session.