iTerm2: Make cursor guide work better for me.
The 'Smart box cursor color' option for the cursor guide wasn't doing it
for me; it changed cursor color while typing (possibly due to the
colored prompt in the line above the cursor guide).

The color of the cursor guide, in particular the opacity setting also
interfered with the simple prompt that I sometimes use; this was due to
the fact that there was no opacity in the cursor guide and the cursor
guide was on the same line as the prompt when using the simple prompt
(`prompt_simple()` that is).

This change incorporates both the removal of the 'Smart box cursor
color' option, as well as adding opacity (and changing color a bit) of
the cursor guide, and some minor changes to the cursor color.