Add function to check for updates in Home Brew.
On one hand blindly upgrading might bring surprises, on the other it'll
_always_ include DropBox (This is always on 'latest'), even when it is
_not_ updated (which is most of the time).

Since DropBox seems to manage its updates pretty well, it seems better
to exclude it from blindly upgrading, which, for now, means manually
checking.  Which is better, I feel, because it allows me to cater for
specif changes (Like Python3 for instance).