Remove iTerm2, using default Terminal.
iTerm2 is a great terminal, but I seem to rely on it less and less.
Also, to be more up to date, I subscribed to the beta channel which
sometimes has bugs/regressions (it's beta after all), but the last one
where I could not mouse-select partial text _really_ annoyed me, so I
temporarily switched to OSX's default terminal, changed some basic
settings and I'm actually happy with it, so why waste the disk space of
having an extra terminal installed?

It might be that I'll miss something iTerm2-ish later down the line, but
it's easily re-installed of course.