Git alias `l` shows 10 lines instead of 15.
I've been changing the way I put my Terminal window, leaving it at the
bottom of the display; this means more columns, but less lines in both
editor (which I now keep at the top of the display) and terminal.
On the laptop screen 15 lines forces a scroll, so I'm reducing it to 10.
I think this still should be enough for a quick look.  We'll see...