Export default LANG.
This fixes the copy paste bork I encountered yanking stuff to nvim's '+'
register.
Don't really know what happened, never encountered this before.
`:language` on vim was reporting correct locale with UTF8, but in Neovim
it was not.  Exporting $LANG this way sets the output of `:language` the
same in both Vim and Neovim.