Make tpl() list only file names for selecting.
The full on expanded $DIR_TEMPLATES provided no useful information when
selecting the correct template, so it had to go.  The substitution in
the shell parameter expansion before doing the substitution with sed was
both necessary and kinda (back)slash confusing, but all in all clear
when given some attention to interpretation.