Add $DIR_TEMPLATES also to the preview command.
In [959c304c] a filter was introduced before feeding FZF.  But because
of the --preview command in FZF and its use of '{}' to represent the
selection, this did not work anymore since it was only fed a file name
and not a full path.  This went unnoticed, because I've tested with
`.gitignore` in the template directory and that file is of course also
present in this repository.