Output of `uli` is copied to clipboard.
Since `uli` just executed `drush uli` on the server, it had
'http://default' in the beginning which is totally useless.  It was a
hassle to manually copy the path part of the uli and then paste it into
the browser's location bar or in a chat.

With this change, the irrelevant 'http://default' prefix is removed from
the output _and_ the output is directly redirected to the clipboard
instead of being sent to STDOUT, making `uli` just that tad more
efficient.