Starship prompt: Initial customization.
Nothing special, some more uniform coloring to detect the prompt in a
long list of commands + outputs in a terminal window.
And also added a success/error symbol  + the error code if non-zero,
which I was getting used to in my custom rolled prompt.

Not too sure about the default colors, but it is well readable.
Not convinced at all about all the language versions in the prompt, but
keeping it for the time being since Python module also shows virtualenv,
upon which I rely a lot.

And last but not least, added a format that is lightweight for
repositories that take a bit of time when doing a `git status`.