Setting an e-mail address is now implicit
Too often, I had a work e-mail address in commits of non-work
repositories.  This happens when working on both work and non-work
repositories from the same computer, since user configuration was stored
globally in `.gitconfig-user` on a machine.

By using the configuration option `user.useConfigOnly` we force to set a
user per repository.  The name can be set, that does not change.
I don't know yet how hard of an inconvenience this will be, but as I
look at it now, I prefer the correct user being set over the possible
frustration of configuring it on a per repository base... Future Self
will be the judge.