Fix Bash completion for the Git `g` function
For some reason, this stopped functioning, complaining that the function
`_git` could not be found... upon searching, I discovered that at some
point in time the sequence of which you set the complete for aliases and
functions for Git should be come _after_ sourcing the Git autocompletion
for Bash, which wasn't the case, but changing order did not help at all.

In an older SO post, I found the reference to the __git_wrap__git_main
function (instead of _git (Yes, I've tried `__git` with the same
effect)) and that seemed to be working, so Good Enough for me for the
time being.