Configuring Kitty terminal, might replace iTerm2
I've been slightly annoyed by the screen (re)draws of iTerm2-beta:
    * Sometimes slower than I would like to
    * Switching window size a lot and then exiting Neovim messes up the
      screen immensely, very annoying
    * FZF in Neovim sometimes refuses to show the preview window and
      outputs garbage on screen.  Luckily this isn't saved in the buffer
      or anything, but it _is_ scary to see.

Came across some recommendation to try Kitty and so I did again; in the
brief run I had so far, it did not garble the screen once and FZF worked
flawlessly.
The only thing I'm finding annoying is that the font rendering is not on
par of that of iTerm2's which is a shame, but it's minimal enough for me
to spend half an hour configuring some basic stuff and deciding to give
it a try.