Make `pussy` alias a function to be able to use `$_`
In [cdafa82] we added the alias, but it did not support fetching the
last argument of the previous command.  By making it a function, we can
leverage the usage of `$_` to do that for us.  Now we can just type
`pussy` after an ssh into a server that hasn't got the correct terminfo
files and be all good and dandy.