Take Python Virtual environments into account if possible
When a virtualenv exists, `project_goto` (renamed from `goto_project` in
line with all the ssh_* functions) will try to activate it.
This is more POC than a robust solution, but does the basic thing it
should do; deactivate the current virtualenv and activate the new one
(if either one is present).
The solution might be done cleaner, for instance now; it will deactivate
and reactivate the virtualenv, even if the one requested to be activated
was already activated before, etc... but it works fine enough for the
time being.

The 1-1 mapping of project name and virtualenv could make some
virtualenvs rather long in name, since at work we prefix this with the
names of the technology in use and what not, but since Python is not
used that much at work, it's fine I guess.  Also, if the virtualenv is
named the same as the project, it eliminates confusion... It _does_ make
a very long prompt though :)

Closes #37