Kitty tab bar config tweaks
The tab bar is now at the top of the window, where it curses a lot less
with Neovim's command line.

Behaviour of closing tabs is more to my liking (the 'previous' strategy
was just confusing to me, closed too many tabs accidentally that way).

Not too sure about including the tab index only in other tabs than the
active one, but we'll see how that one goes.