Training template is now 5 week iteration
Going with a 4 week training + 1 week rest/project/recuperation week as
one 5 week iteration.
Further changes include: 4 lines per day as to visually distinguish
between different parts of the day; a morning pinch session on Monday
could go with a boulder session on Tuesday evening for example.
A column was added in the beginning to count the weeks and have a place
to fill in dates, making it a bit easier to plan/compare with a
calendar.